backend/pyproject.toml
[project]
name = "fridgeflow-backend"
version = "0.1.0"
readme = "README.md"
requires-python = ">=3.10,<3.12"
dependencies = [
    "colorlog>=6.9.0",
    "django-cors-headers>=4.6.0",
    "django-environ>=0.11.2",
    "django-extensions>=3.2.3",
    "django-guid>=3.5.0",
    "django-health-check>=3.18.3",
    "django-ninja>=1.3.0",
    "gunicorn>=23.0.0",
    "httpx>=0.28.1",
    "psycopg2-binary>=2.9.10",
    "python-json-logger>=3.2.1",
    "RapidFuzz>=3.11.0"
]

[dependency-groups]
dev = [
    "django-debug-toolbar>=4.4.6",
    "ruff>=0.8.4",
]

[tool.ruff]
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    "__pypackages__",
    "_build",
    "buck-out",
    "dist",
    "node_modules",
    "venv",
    ".vscode",
    "migrations",
]
line-length = 79

[tool.ruff.lint]
select = ["ALL"]
ignore = ["D", "ANN", "ARG", "COM812", "ISC001", "PT009"]

[tool.ruff.format]
docstring-code-format = true
docstring-code-line-length = 79

indent-style = "space"
quote-style = "double"
skip-magic-trailing-comma = false

[tool.ruff.lint.pylint]
max-args = 6
